Added three right click hooks to modifiers * onItemRightClick * onItemUseFirst * itemInteractionForEntity
Changed existing right click logic on tools to hidden ability modifiers
* Mattock, Excavator: Shovel Transform (paths)
* Hand Axe, Broad Axe: Axe Transform (log stripping)
* Mattock, Excavator: Hoe Transform (hoe), Shears, Kama (harvesting)
