+++ src/main/java/slimeknights/tconstruct/tools/ToolDefinitions.java
@@ -12,6 +12,8 @@ import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import com.google.common.collect.ImmutableList;
+
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class ToolDefinitions {
   // rock
@@ -28,26 +30,36 @@ public final class ToolDefinitions {
   // dirt
   public static final ToolDefinition MATTOCK = new ToolDefinition(
     ToolBaseStatDefinitions.MATTOCK,
-    requirements(TinkerToolParts.smallAxeHead, TinkerToolParts.toolHandle, TinkerToolParts.pickaxeHead));
+    requirements(TinkerToolParts.smallAxeHead, TinkerToolParts.toolHandle, TinkerToolParts.pickaxeHead),
+    () -> Collections.singletonList(new ModifierEntry(TinkerModifiers.shovelTransform.get(), 1)));
   public static final ToolDefinition EXCAVATOR = new ToolDefinition(
     ToolBaseStatDefinitions.EXCAVATOR,
-    requirements(TinkerToolParts.largePlate, TinkerToolParts.toughHandle, TinkerToolParts.largePlate, TinkerToolParts.toughHandle));
+    requirements(TinkerToolParts.largePlate, TinkerToolParts.toughHandle, TinkerToolParts.largePlate, TinkerToolParts.toughHandle),
+    () -> Collections.singletonList(new ModifierEntry(TinkerModifiers.shovelTransform.get(), 1)));
 
   // wood
   public static final ToolDefinition HAND_AXE = new ToolDefinition(
     ToolBaseStatDefinitions.HAND_AXE,
-    requirements(TinkerToolParts.smallAxeHead, TinkerToolParts.toolHandle, TinkerToolParts.toolBinding));
+    requirements(TinkerToolParts.smallAxeHead, TinkerToolParts.toolHandle, TinkerToolParts.toolBinding),
+    () -> Collections.singletonList(new ModifierEntry(TinkerModifiers.axeTransform.get(), 1)));
   public static final ToolDefinition BROAD_AXE = new ToolDefinition(
     ToolBaseStatDefinitions.BROAD_AXE,
-    requirements(TinkerToolParts.broadAxeHead, TinkerToolParts.toughHandle, TinkerToolParts.pickaxeHead, TinkerToolParts.largePlate));
+    requirements(TinkerToolParts.broadAxeHead, TinkerToolParts.toughHandle, TinkerToolParts.pickaxeHead, TinkerToolParts.largePlate),
+    () -> Collections.singletonList(new ModifierEntry(TinkerModifiers.axeTransform.get(), 1)));
 
   // scythes
   public static final ToolDefinition KAMA = new ToolDefinition(
     ToolBaseStatDefinitions.KAMA,
-    requirements(TinkerToolParts.swordBlade, TinkerToolParts.toolHandle, TinkerToolParts.toolBinding));
+    requirements(TinkerToolParts.swordBlade, TinkerToolParts.toolHandle, TinkerToolParts.toolBinding),
+    () -> ImmutableList.of(new ModifierEntry(TinkerModifiers.hoeTransform.get(), 1),
+                           new ModifierEntry(TinkerModifiers.shears.get(), 1),
+                           new ModifierEntry(TinkerModifiers.kama.get(), 1)));
   public static final ToolDefinition SCYTHE = new ToolDefinition(
     ToolBaseStatDefinitions.SCYTHE,
-    requirements(TinkerToolParts.broadBlade, TinkerToolParts.toughHandle, TinkerToolParts.toolBinding, TinkerToolParts.toughHandle));
+    requirements(TinkerToolParts.broadBlade, TinkerToolParts.toughHandle, TinkerToolParts.toolBinding, TinkerToolParts.toughHandle),
+    () -> ImmutableList.of(new ModifierEntry(TinkerModifiers.hoeTransform.get(), 1),
+                           new ModifierEntry(TinkerModifiers.shears.get(), 1),
+                           new ModifierEntry(TinkerModifiers.kama.get(), 1)));
 
   // swords
   public static final ToolDefinition BROADSWORD = new ToolDefinition(
