+++ src/main/java/slimeknights/tconstruct/tools/harvest/MattockTool.java
@@ -4,9 +4,6 @@ import com.google.common.collect.Sets;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.material.Material;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.ItemUseContext;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.SoundEvents;
 import net.minecraftforge.common.ToolType;
 import slimeknights.tconstruct.library.tools.ToolDefinition;
 import slimeknights.tconstruct.library.tools.helper.ToolHarvestLogic;
@@ -53,10 +50,5 @@ public class MattockTool extends ToolCore {
   @Override
   public ToolHarvestLogic getToolHarvestLogic() {
     return HARVEST_LOGIC;
-  }
-
-  @Override
-  public ActionResultType onItemUse(ItemUseContext context) {
-    return getToolHarvestLogic().transformBlocks(context, ToolType.SHOVEL, SoundEvents.ITEM_SHOVEL_FLATTEN, true);
   }
 }
