+++ src/main/java/slimeknights/tconstruct/tools/modifiers/ability/ShearsAbilityModifier.java
@@ -1,0 +1,115 @@
+package slimeknights.tconstruct.tools.modifiers.ability;
+
+import java.util.List;
+import java.util.Random;
+
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.enchantment.Enchantments;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.item.ItemEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Hand;
+import net.minecraft.world.World;
+import net.minecraftforge.common.IForgeShearable;
+import slimeknights.tconstruct.library.modifiers.SingleUseModifier;
+import slimeknights.tconstruct.library.tools.helper.ToolDamageUtil;
+import slimeknights.tconstruct.library.tools.item.ToolCore;
+import slimeknights.tconstruct.library.tools.nbt.ToolStack;
+
+public class ShearsAbilityModifier extends SingleUseModifier {
+  private final int priority;
+  
+  public ShearsAbilityModifier(int color, int priority) {
+    super(color);
+    this.priority = priority;
+  }
+  
+  public int getPriority() {
+    return priority;
+  }
+  
+  public boolean shouldDisplay(boolean advanced) {
+    return priority > Short.MIN_VALUE;
+  }
+  
+  /**
+   * Swings the given's player hand
+   *
+   * @param player the current player
+   * @param hand the given hand the tool is in
+   */
+  protected void swingTool(PlayerEntity player, Hand hand) {
+    player.swingArm(hand);
+    player.spawnSweepParticles();
+  }
+
+  private boolean isShears(ToolStack tool) {
+    Item item = tool.getItem();
+    if (item instanceof ToolCore)
+      {
+      ToolCore core = (ToolCore) item;
+      return core.isShears(tool);
+      }
+    // We're on a tool that isn't ToolCore and wants to be shears, okay, cool.
+    return true;
+  }
+  
+  @Override
+  public ActionResultType itemInteractionForEntity(ToolStack tool, int level, ItemStack stack, PlayerEntity playerIn, LivingEntity target, Hand hand) {
+    // only run AOE on shearable entities
+    if (isShears(tool) && target instanceof IForgeShearable) {
+      int fortune = EnchantmentHelper.getEnchantmentLevel(Enchantments.FORTUNE, stack);
+
+      if (!tool.isBroken() && this.shearEntity(stack, playerIn.getEntityWorld(), playerIn, target, fortune)) {
+        ToolDamageUtil.damageAnimated(tool, 1, playerIn, hand == Hand.MAIN_HAND ? EquipmentSlotType.MAINHAND : EquipmentSlotType.OFFHAND);
+        this.swingTool(playerIn, hand);
+        return ActionResultType.SUCCESS;
+      }
+    }
+
+    return ActionResultType.PASS;
+  }
+
+  /**
+   * Tries to shear an given entity, returns false if it fails and true if it succeeds
+   *
+   * @param itemStack the current item stack
+   * @param world the current world
+   * @param playerEntity the current player
+   * @param entity the entity to try to shear
+   * @param fortune the fortune to apply to the sheared entity
+   * @return if the sheering of the entity was performed or not
+   */
+  private boolean shearEntity(ItemStack itemStack, World world, PlayerEntity playerEntity, Entity entity, int fortune) {
+    if (!(entity instanceof IForgeShearable)) {
+      return false;
+    }
+
+    IForgeShearable target = (IForgeShearable) entity;
+
+    if (target.isShearable(itemStack, world, entity.getPosition())) {
+      if (!world.isRemote) {
+        List<ItemStack> drops = target.onSheared(playerEntity, itemStack, world, entity.getPosition(), fortune);
+        Random rand = world.rand;
+
+        drops.forEach(d -> {
+          ItemEntity ent = entity.entityDropItem(d, 1.0F);
+
+          if (ent != null) {
+            ent.setMotion(ent.getMotion().add((rand.nextFloat() - rand.nextFloat()) * 0.1F, rand.nextFloat() * 0.05F, (rand.nextFloat() - rand.nextFloat()) * 0.1F));
+          }
+        });
+      }
+      return true;
+    }
+
+    return false;
+  }
+  
+  
+}
